+++ Runtime/Tweens/Components/InfoComponents/TweenScale.cs
@@ -1,8 +1,10 @@
 ﻿using Unity.Entities;
 using Unity.Mathematics;
+using Unity.Transforms;
 
 namespace Timespawn.EntityTween.Tweens
 {
+    [WriteGroup(typeof(Scale))]
     public struct TweenScale : IComponentData, ITweenId, ITweenInfo<float3>
     {
         public int Id;