+++ Tests/Runtime/Tweens/DestroyTests.cs
@@ -1,14 +1,21 @@
 ﻿using NUnit.Framework;
 using Timespawn.EntityTween.Tweens;
 using Unity.Entities;
-using Unity.Mathematics;
+
+#if UNITY_TINY_ALL_0_31_0
+using Unity.Tiny;
+#endif
+
+#if UNITY_2D_ENTITIES
+using Unity.U2D.Entities;
+#endif
 
 namespace Timespawn.EntityTween.Tests.Tweens
 {
     public class DestoryTests : TweenTestFixture
     {
         [Test]
-        public void Translation()
+        public void Move()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Move(EntityManager, entity, TestStartFloat3, TestEndFloat3, TestDuration);
@@ -16,7 +23,7 @@ namespace Timespawn.EntityTween.Tests.Tweens
         }
 
         [Test]
-        public void Rotation()
+        public void Rotate()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Rotate(EntityManager, entity, TestStartQuat, TestEndQuat, TestDuration);
@@ -30,6 +37,18 @@ namespace Timespawn.EntityTween.Tests.Tweens
             Tween.Scale(EntityManager, entity, TestStartFloat3, TestEndFloat3, TestDuration);
             Test<TweenScale>(entity);
         }
+
+#if UNITY_TINY_ALL_0_31_0 || UNITY_2D_ENTITIES
+    
+        [Test]
+        public void Tint()
+        {
+            Entity entity = EntityManager.CreateEntity();
+            Tween.Tint(EntityManager, entity, TestStartFloat4, TestEndFloat4, TestDuration);
+            Test<TweenTint>(entity);
+        }
+
+#endif
 
         private void Test<TTweenInfo>(Entity entity)
             where TTweenInfo : struct, IComponentData